[Templates] - Upped DesktopGL template from .NET 4.5.2 to 4.6 [Redgrin Plugin] - Added additional info to the Redgrin Screen UI to indicate where to put the NetworkManager initialize code. - Fixed bug where random port number could be generates as negative based on the project name hash. [Glue] - Added support for AssetTypeInfos to have a Func for the custom load code so it can have conditional logic rather than always being the same. - Added controls for Glue resolution behavior on resize. - Plugins can be drag+dropped on the tree view window to install them. [Glue + Gum] - Added fixed aspect ratio functionality to Glue + Gum plugin. More to do here [AnimationEditor Plugin] - CTRL+click when snap to grid will now add new frames, just like CTRL click on magic wand. - Added Alt+arrow reordering of frames - Added right-click reorder menu items for frames. [Engine] - Added docs to TileNodeNetwork. [Glue + Tiled] Added docs to MapDrawableBatch.

+++ FRBDK/Glue/RedGrinPlugin/CodeGenerators/NetworkConfigurationCodeGenerator.cs
@@ -38,7 +38,7 @@ namespace RedGrinPlugin.CodeGenerators
             var constructor = codeBlock.Constructor("public", "GameNetworkConfiguration", "");
 
 
-            var portNumber = projectName.GetHashCode() % (ushort.MaxValue - 1024) + 1024;
+            var portNumber = (ushort)( (ushort)(projectName.GetHashCode() % (ushort.MaxValue - 1024)) + (ushort)1024);
             
 
             constructor.Line($"ApplicationName = \"{projectName}\";");
